+++ src/python/fm/web/FileMoverService.py
@@ -7,10 +7,11 @@ FileMover web-server
 
 # system modules
 import os
+import re
+import cgi
 import sys
 import stat
 import time
-import urllib
 import traceback
 
 # CherryPy/Cheetah modules
@@ -22,14 +23,24 @@ from   cherrypy import config as cherryconf
 # FileMover modules
 from   fm.utils.FMWSConfig  import fm_config
 from   fm.core.FileManager import FileManager, validate_lfn
-from   fm.core.Status import StatusCode
+from   fm.core.Status import StatusCode, StatusMsg
 from   fm.dbs.DBSInteraction import DBS, dbsinstances
-from   fm.utils.Utils import sizeFormat, getArg
+from   fm.utils.Utils import sizeFormat
 
 # WMCore/WebTools modules
 from WMCore.WebTools.Page import TemplatedPage
 
 CODES = {'valid': 0, 'too_many_request': 1, 'too_many_lfns': 2}
+
+def removeLfn(lfn):
+    """HTML snippet for LFN remove request"""
+    msg = "<a href=\"javascript:ajaxRemove('%s')\">Remove</a>" % lfn
+    return msg
+
+def cancelLfn(lfn):
+    """HTML snippet for LFN cancel request"""
+    msg = "<a href=\"javascript:ajaxCancel('%s')\">Cancel</a>" % lfn
+    return msg
 
 def checkarg(kwds, arg, atype=str):
     """Check arg in a dict that it has provided type"""
@@ -52,10 +63,12 @@ def checkargs(func):
         if  keys:
             raise HTTPError(500, 'Unsupported key')
         if  checkarg(kwds, 'lfn'):
+            lfn = kwds.get('lfn').strip()
             try:
                 validate_lfn(kwds.get('lfn'))
             except:
                 raise HTTPError(500, 'Unsupported LFN')
+            kwds['lfn'] = lfn
         if  checkarg(kwds, 'dataset'):
             pat = re.compile('/.*/.*/.*')
             if  not pat.match(kwds.get('dataset')):
@@ -128,6 +141,12 @@ def spanId(lfn):
     """assign id for span tag based on provided lfn"""
     return lfn.split("/")[-1].replace(".root","")
 
+def handleExc():
+    """exception handler"""
+    traceback.print_exc()
+    page  = "<div>Request failed</div>"
+    return page
+
 class FileMoverService(TemplatedPage):
     """FileMover web-server based on CherryPy"""
     def __init__(self, config):
@@ -181,37 +200,12 @@ class FileMoverService(TemplatedPage):
                            'tools.gzip.mime_types': mime_types,
                           })
 
-    def handleExc(self):
-        """exception handler"""
-        traceback.print_exc()
-        page  = "<div>Request failed</div>"
-        return page
-
-#    @expose
-#    @tools.secmodv2()
-#    def index(self):
-#        """default service method"""
-#        page = self.getTopHTML()
-#        user = cherrypy.request.user['dn']
-#        print "\n### USER", user
-#        self.addUser(user)
-#        page += self.userForm(user)
-#        page += getBottomHTML()
-#        return page
-
     @expose
-#    @tools.secmodv2()
-    def getUserName(self):
-        """Get userName from cherrypy header"""
-#        return cherrypy.request.user['dn']
-        # TODO: need to get DN from FrontEndAuth module.
-        return "test"
-
-    @expose
+    @tools.secmodv2()
     def index(self):
         """default service method"""
         page = self.getTopHTML()
-        user = self.getUserName()
+        user = cherrypy.request.user['name']
         self.addUser(user)
         page += self.userForm(user)
         page += getBottomHTML()
@@ -223,7 +217,16 @@ class FileMoverService(TemplatedPage):
             self.userDict[user] = ([], [])
 
     def makedir(self, user):
-        """Create user dir structure on a server"""
+        """
+        Create user dir structure on a server. It has the following structure
+        <path>
+        ├── download
+        │   ├── <user>
+        │   │   └── softlinks
+        The download/<user> area are used for storage of hardlinks to files in
+        FileMover pool (for bookkeeping purposes), while softlinks are used to
+        refer to actual location of the files in a pool.
+        """
         try:
             os.makedirs("%s/%s/softlinks" % (self.download_dir, user))
         except Exception as _exc:
@@ -234,44 +237,75 @@ class FileMoverService(TemplatedPage):
         page = self.templatepage('templateTop', url=self.url)
         return page
 
-    def LfnRows(self, iList, sList):
-        """HTML formatted list of LNFs"""
+    def updateUserPage(self, user):
+        """
+        Update user HTML page with current status of all LFNs
+        which belong to that user.
+        """
+        try:
+            iList, sList = self.userDict[user]
+        except Exception as _exc:
+            return ""
         page  = ""
         style = ""
         for idx in xrange(0, len(iList)):
-            lfn   = iList[idx]
-            fstat = sList[idx]
+            lfn = iList[idx]
+            try:
+                statusCode, statusMsg = sList[idx]
+            except:
+                print "\n### userDict", self.userDict[user]
+                traceback.print_exc()
+                HTTPError(500, "Server error")
+            if  statusCode == StatusCode.DONE: # append remove request
+                statusMsg += " | " + removeLfn(lfn)
+            else: # sanitize msg and append cancel request
+                statusMsg  = cgi.escape(statusMsg) + " | " + cancelLfn(lfn)
             if  style:
                 style = ""
             else:
                 style = "zebra"
             spanid = spanId(lfn)
             page += self.templatepage('templateLfnRow', style=style, lfn=lfn, \
-                spanid=spanid, fstat=fstat)
+                spanid=spanid, statusCode=statusCode, statusMsg=statusMsg)
         return page
 
-    def getStat(self, user):
-        """get status of current user's job"""
-        try:
-            iList, sList = self.userDict[user]
-        except Exception as _exc:
-            return ""
-        page = self.LfnRows(iList, sList)
-        return page
-        
-    def getStatForLfn(self, user, lfn):
-        """return status of user requested lfn"""
-        validate_lfn(lfn)
+    def updatePageWithLfnInfo(self, user, lfn):
+        """Update page with LFN info"""
         page  = ""
         lfnList, statList = self.userDict[user]
+        self.makedir(user)
         try:
-            idx    = lfnList.index(lfn)
-            fstat  = statList[idx]
-            page  += """%s""" % fstat
-            if  fstat.find("Download") == -1:
-                page += " | <a href=\"javascript:ajaxCancel('%s')\">Cancel</a>"\
-                         % lfn
+            idx = lfnList.index(lfn)
+            statusCode, statusMsg  = statList[idx]
+            if  statusCode == StatusCode.DONE:
+                filename = lfn.split('/')[-1]
+                pfn      = os.path.join(self.transfer_dir, lfn[1:])
+                if  os.path.isfile(pfn):
+                    if  not os.path.isfile("%s/%s/%s" \
+                        % (self.download_dir, user, filename)):
+                        try:
+                            os.link(pfn, "%s/%s/%s" \
+                                % (self.download_dir, user, filename))
+                            os.symlink(pfn, "%s/%s/softlinks/%s" \
+                                % (self.download_dir, user, filename))
+                        except Exception as _exc:
+                            traceback.print_exc()
+                    link     = "download/%s/%s" % (user, filename)
+                    filepath = "%s/%s/%s" % (self.download_dir, user, filename)
+                    fileStat = os.stat(filepath)
+                    fileSize = sizeFormat(fileStat[stat.ST_SIZE])
+                    msg  = "<a href=\"%s/%s\">Download (%s)</a> |" \
+                                % (self.url, link, fileSize)
+                    statList[idx] = (StatusCode.DONE, msg)
+                else:
+                    statList[idx] = (StatusCode.TRANSFER_STATUS_UNKNOWN, \
+                        StatusMsg.TRANSFER_STATUS_UNKNOWN)
+                    msg = cgi.escape(StatusMsg.TRANSFER_STATUS_UNKNOWN)
+                page += msg + " | " + removeLfn(lfn)
+            else:
+                page += cgi.escape(statusMsg) + " | " + cancelLfn(lfn)
         except Exception as _exc:
+            traceback.print_exc()
             print lfn
             print self.userDict
         return page
@@ -307,18 +341,8 @@ class FileMoverService(TemplatedPage):
         page = self.ajaxResponse(page)
         return page
         
-    def nActiveDownloads(self, user):
-        """check number of active downloads/user"""
-        _lfnList, statList = self.userDict[user]
-        count = 0
-        for istat in statList:
-            if  istat.find("Download") == -1:
-                count += 1
-        return  count
-        
     def addLfn(self, user, lfn):
         """add LFN request to the queue"""
-        validate_lfn(lfn)
         # check how may requests in total user placed today
         today = time.strftime("%Y%m%d", time.gmtime(time.time()))
         if  self.userDictPerDay.has_key(user):
@@ -335,7 +359,8 @@ class FileMoverService(TemplatedPage):
         if  len(lfnList) < int(self.max_transfer):
             if  not lfnList.count(lfn):
                 lfnList.append(lfn)
-                statList.append('requested')
+                status = (StatusCode.REQUESTED, StatusMsg.REQUESTED)
+                statList.append(status)
                 nRequests, day = self.userDictPerDay[user]
                 self.userDictPerDay[user] = (nRequests+1, today)
             else:
@@ -347,12 +372,12 @@ class FileMoverService(TemplatedPage):
         
     def delLfn(self, user, lfn):
         """delete LFN from the queue"""
-        validate_lfn(lfn)
         lfnList, statList = self.userDict[user]
         if  lfnList.count(lfn):
             idx = lfnList.index(lfn)
             lfnList.remove(lfn)
             statList.remove(statList[idx])
+            self.userDict[user] = (lfnList, statList)
             self._remove(user, lfn)
 
     def checkUserCache(self, user):
@@ -375,79 +400,37 @@ class FileMoverService(TemplatedPage):
             fileSize = sizeFormat(fileStat[stat.ST_SIZE])
             link     = "download/%s/%s" % (user, ifile)
             lfn      = abspath.replace(self.transfer_dir, "")
-            msg  = "<a href=\"%s/%s\">Download (%s)</a> |" \
+            msg  = "<a href=\"%s/%s\">Download (%s)</a> " \
                         % (self.url, link, fileSize)
-            msg += "<a href=\"javascript:ajaxRemove('%s')\">Remove</a> " \
-                        % lfn
             if  not lfnList.count(lfn):
                 lfnList.append("%s" % lfn)
-                statList.append(msg)
+                status = (StatusCode.DONE, msg)
+                statList.append(status)
         self.userDict[user] = (lfnList, statList)
-        page += self.getStat(user)
+        page += self.updateUserPage(user)
         return page
         
     def setStat(self, user, _lfn=None):
-        """update status of retrieving LFN"""
-        statCode = -1
+        """
+        Update status of LFN in transfer for given user and return
+        FileManager status code.
+        """
+        statCode = StatusCode.UNKNOWN
         self.addUser(user)
-        self.makedir(user)
         lfnList, statList = self.userDict[user]
         for idx in xrange(0, len(lfnList)):
             lfn    = lfnList[idx]
             if  _lfn and lfn != _lfn:
                 continue
-            prevStat = statList[idx]
             status = self.fmgr.status(lfn)
-            if  status == "This LFN has not been requested yet!":
-                status = (0, 'orphan')
-            elif status == 'Unknown status; internal error.':
-                status = (0, 'orphan')
-            if  lfn.find("/") == -1:
-                statCode = 0
-                statMsg  = prevStat 
-                statList[idx] = statMsg
-                ifile = lfn
-                pfn = "%s/%s/%s" % (self.download_dir, user, ifile)
-            else:
-                statCode = status[0]
-                statMsg  = status[1]
-                statList[idx] = statMsg
-                iList  = lfn.split("/")
-                ifile  = iList[-1]
-                idir   = self.transfer_dir + '/'.join(iList[:-1])
-                pfn    = os.path.join(idir, ifile)
-            if  statCode == 0:
-                if  os.path.isfile(pfn):
-                    if  not os.path.isfile("%s/%s/%s" \
-                        % (self.download_dir, user, ifile)):
-                        try:
-                            os.link(pfn, "%s/%s/%s" \
-                                % (self.download_dir, user, ifile))
-                            os.symlink(pfn, "%s/%s/softlinks/%s" \
-                                % (self.download_dir, user, ifile))
-                        except Exception as _exc:
-                            traceback.print_exc()
-                    link     = "download/%s/%s" % (user, ifile)
-                    filepath = "%s/%s/%s" % (self.download_dir, user, ifile)
-                    fileStat = os.stat(filepath)
-                    fileSize = sizeFormat(fileStat[stat.ST_SIZE])
-                    lfn      = pfn.replace(self.transfer_dir, "")
-                    msg  = "<a href=\"%s/%s\">Download (%s)</a> |" \
-                                % (self.url, link, fileSize)
-                    msg += "<a href=\"javascript:ajaxRemove('%s')\">Remove</a>"\
-                         % lfn
-                    statList[idx] = msg
-                else:
-                    msg  = "Unable to transfer"
-                    msg += "<a href=\"javascript:ajaxRemove"
-                    msg += "('%s')\">Remove</a> " % lfn
-                    statList[idx] = msg
+            statList[idx] = status
+            statCode = status[0]
+            break
         self.userDict[user] = (lfnList, statList)
         return statCode
 
     def _remove(self, user, lfn):
         """remove requested LFN from the queue"""
-        validate_lfn(lfn)
         ifile = lfn.split("/")[-1]
         userdir = "%s/%s" % (self.download_dir, user)
         # remove soft-link from user download area
@@ -476,17 +459,17 @@ class FileMoverService(TemplatedPage):
             pass
 
     @expose
+    @tools.secmodv2()
     @checkargs
-    def remove(self, lfn, **kwargs):
+    def remove(self, lfn, **_kwargs):
         """remove requested LFN from the queue"""
-        validate_lfn(lfn)
-        user = self.getUserName()
-        force_remove = getArg(kwargs, 'force_remove', 0)
-        page = ""
-        self._remove(user, lfn)
-        if  not force_remove:
-            self.delLfn(user, lfn)
-            page += self.getStat(user)
+        user = cherrypy.request.user['name']
+        self.delLfn(user, lfn)
+        page = self.updateUserPage(user)
+        try:
+            self.fmgr.cancel(lfn)
+        except Exception as _exc:
+            page = handleExc()
         return self.ajaxResponse(page)
 
     def tooManyRequests(self, user):
@@ -501,18 +484,16 @@ class FileMoverService(TemplatedPage):
                 page = self.templatepage('templateTooManyRequests', \
                     max_transfer=self.max_transfer, \
                     day_transfer=self.day_transfer, fulldesc=True)
-        page += self.getStat(user)
+        page += self.updateUserPage(user)
         return self.ajaxResponse(page)
 
     @expose
+    @tools.secmodv2()
     @checkargs
     def request(self, lfn, **kwargs):
         """place LFN request"""
-        user = self.getUserName()
-        lfn = lfn.strip() # remove spaces around lfn
-        lfn = urllib.unquote(lfn)
-        validate_lfn(lfn)
-        page = getArg(kwargs, 'page', '')
+        user = cherrypy.request.user['name']
+        page = kwargs.get('page')
         lfnStatus = self.addLfn(user, lfn)
         if  not lfnStatus:
             return self.tooManyRequests(user)
@@ -525,28 +506,31 @@ class FileMoverService(TemplatedPage):
             else:
                 page += self.templatepage('templateRequest', \
                         lfn=lfn, msg='in_queue')
-            page += self.getStat(user)
+            page += self.updateUserPage(user)
+            page += self.templatepage('templateTimeout', \
+                            fun="ajaxStatusOne", req=lfn, msec=1000)
         except Exception as _exc:
-            page = self.handleExc()
+            page = handleExc()
+#        print "\n### new page\n"
+#        print page
         page = self.ajaxResponse(page)
         return page
         
     @expose
+    @tools.secmodv2()
     @checkargs
     def cancel(self, lfn, **_kwargs):
         """cancel LFN request"""
-        user = self.getUserName()
-        lfn = urllib.unquote(lfn)
-        validate_lfn(lfn)
+        user = cherrypy.request.user['name']
         self.delLfn(user, lfn)
         page = ""
         try:
             self.fmgr.cancel(lfn)
             page = self.templatepage('templateRequest', \
-                        lfn=lfn, msg='discarded')
+                            lfn=lfn, msg='discarded')
         except Exception as _exc:
-            page = self.handleExc()
-        page += self.getStat(user)
+            page = handleExc()
+        page += self.updateUserPage(user)
         page  = self.ajaxResponse(page)
         return page
 
@@ -578,13 +562,11 @@ class FileMoverService(TemplatedPage):
         return page
 
     @expose
+    @tools.secmodv2()
     @checkargs
     def statusOne(self, lfn, **_kwargs):
         """return status of requested LFN"""
-        user = self.getUserName()
-        lfn  = urllib.unquote(lfn)
-        lfn  = lfn.strip()
-        validate_lfn(lfn)
+        user = cherrypy.request.user['name']
         page = ""
         spanid = spanId(lfn)
         page += """<span id="%s" name="%s">""" % (spanid, spanid)
@@ -592,12 +574,12 @@ class FileMoverService(TemplatedPage):
             statCode = self.setStat(user, lfn)
             if  statCode and statCode != StatusCode.TRANSFER_FAILED:
                 page += self.templatepage('templateLoading', msg="")
-            page += self.getStatForLfn(user, lfn)
+            page += self.updatePageWithLfnInfo(user, lfn)
             if  statCode and statCode != StatusCode.TRANSFER_FAILED:
                 page += self.templatepage('templateTimeout', \
                                 fun="ajaxStatusOne", req=lfn, msec=1000)
         except Exception as _exc:
-            page = self.handleExc()
+            page += handleExc()
         page += "</span>"
         page = self.ajaxResponse(page, spanId(lfn))
         return page